ams-10812 reorg to fix FLBPluginInit test
- create an interface through which we can call NewClient for storage
- implement the new interface for prod
- use the new interface in test and prod
- move all test implementation stubs into stubs_for_test.go
